1차 구현 완료

+++ src/main/java/project/kingstagram/repository/FollowRepository.java
@@ -36,12 +36,12 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
             value = "select count(1) from Follow f " +
                     " where f.fromUser.userId = :userId"
     )
-    Integer findFollowingCountByUserId(@Param("userId") Long userId);
+    Long findFollowingCountByUserId(@Param("userId") Long userId);
 
     @Query(
             value = "select count(1) from Follow f " +
                     " where f.toUser.userId = :userId"
     )
-    Integer findFollowerCountByUserId(@Param("userId") Long userId);
+    Long findFollowerCountByUserId(@Param("userId") Long userId);
 
 }
